+++ doc/changelog.rst
@@ -23,6 +23,12 @@ Highlights include:
   :meth:`~pymongo.database.Database.list_collection_names`.
 - Support for mongodb+srv:// URIs. See
   :class:`~pymongo.mongo_client.MongoClient` for details.
+- Index management helpers (
+  :meth:`~pymongo.collection.Collection.create_index`,
+  :meth:`~pymongo.collection.Collection.create_indexes`,
+  :meth:`~pymongo.collection.Collection.drop_index`,
+  :meth:`~pymongo.collection.Collection.drop_indexes`,
+  :meth:`~pymongo.collection.Collection.reindex`) now support maxTimeMS.
 
 Deprecations: